--- a/common/pch/app-pch.h
+++ b/common/pch/app-pch.h
@@ -1,0 +1,7 @@
+#include "../types.h"
+#include "../database.h"
+#include "../strings.h"
+#include "../eqemu_logsys.h"
+#include "../eqemu_logsys_log_aliases.h"
+#include "../features.h"
+#include "../global_define.h"