--- a/zone/CMakeLists.txt
+++ b/zone/CMakeLists.txt
@@ -480,12 +480,14 @@ ADD_EXECUTABLE(zone ${zone_sources} ${zone_headers})
 # binary output directory
 INSTALL(TARGETS zone RUNTIME DESTINATION ${CMAKE_INSTALL_PREFIX}/bin)
 
-# precompile headers
+# precompiled headers
 IF (WIN32 AND EQEMU_BUILD_PCH)
-    TARGET_PRECOMPILE_HEADERS(zone PRIVATE ../common/pch/pch.h)
-    TARGET_PRECOMPILE_HEADERS(zone PRIVATE ../common/types.h ../common/eqemu_logsys.h ../common/eqemu_logsys_log_aliases.h ../common/features.h ../common/global_define.h)
-    TARGET_PRECOMPILE_HEADERS(zone PRIVATE mob.h npc.h corpse.h doors.h bot.h entity.h client.h zone.h)
+    TARGET_PRECOMPILE_HEADERS(zone PRIVATE ../common/pch/std-pch.h)
 ENDIF()
+
+# precompiled headers
+TARGET_PRECOMPILE_HEADERS(zone PRIVATE ../common/pch/app-pch.h)
+TARGET_PRECOMPILE_HEADERS(zone PRIVATE merc.h mob.h npc.h corpse.h doors.h bot.h entity.h client.h zone.h)
 
 ADD_DEFINITIONS(-DZONE)