… common/CMakeLists.txt
… common/pch/std-pch.h